+++ R/links.R
@@ -169,19 +169,16 @@ ConnectionsToLinks <- function(conns, ccans = NULL, threshold = 0) {
 #' @param keep.all Retain all peaks in the output. Useful for knowing what peaks
 #' had 0 coefficient and what were not tested due to expression or accessibility
 #' thresholds.
-#' @param enet Use elastic net.
 #' @param genes.use Genes to test. If NULL, determine from expression assay.
 #' @param family Response type (for gene expression values). Valid values are
 #' "gaussian", "poisson", "binomial", or a \code{\link[stats]{glm}}-family
 #' object
-#' @param null Compute null distribution of coefficients using randomly selected
-#' peaks from other chromosomes.
 #' @param n_sample Number of peaks to sample at random when computing the null
 #' distribution.
 #' @param verbose Display messages
 #'
 #' @importFrom Seurat GetAssayData
-#' @importFrom glmnet cv.glmnet
+#' @importFrom glmnet cv.glmnet glmnet
 #' @importFrom stats coef
 #' @importFrom Matrix sparseMatrix rowSums
 #' @importFrom future.apply future_lapply
@@ -201,9 +198,11 @@ LinkPeaks <- function(
   expression.slot = "data",
   gene.coords = NULL,
   binary = FALSE,
+  alpha = 0.95,
   distance = 5e+05,
   min.cells = 10,
   expression.threshold = 0.1,
+  family = "poisson",
   genes.use = NULL,
   keep.all = FALSE,
   n_sample = 300,
@@ -217,9 +216,20 @@ LinkPeaks <- function(
   meta.features <- GetAssayData(
     object = object, assay = peak.assay, slot = "meta.features"
   )
+  features.match <- c("GC.percent", "count")
+  if (!("GC.percent" %in% colnames(x = meta.features))) {
+    stop("GC content per peak has not been computed.\n",
+         "Run RegionsStats before calling this function.")
+  }
   peak.data <- GetAssayData(
     object = object, assay = peak.assay, slot = 'counts'
   )
+  if (!("count" %in% colnames(x = meta.features))) {
+    # compute total count
+    hvf.info <- FindTopFeatures(object = peak.data)
+    hvf.info <- hvf.info[rownames(x = meta.features), ]
+    meta.features <- cbind(meta.features, hvf.info)
+  }
   expression.data <- GetAssayData(
     object = object, assay = expression.assay, slot = expression.slot
   )
@@ -277,52 +287,71 @@ LinkPeaks <- function(
       gene.expression <- expression.data[genes.use[[i]], ]
       gene.chrom <- as.character(x = seqnames(x = gene.coords.use[i]))
 
-      if (sum(peak.use) == 0) {
+      if (sum(peak.use) < 2) {
+        # no peaks close to gene
         return(list("gene" = NULL, "coef" = NULL, "pval" = NULL))
       } else {
         peak.access <- t(x = peak.data[peak.use, ])
-        coef.result <- cor(
-          x = as.matrix(x = peak.access),
-          y = as.matrix(x = gene.expression),
-          method = "spearman"
+        cvfit <- cv.glmnet(
+          x = peak.access,
+          y = gene.expression,
+          alpha = alpha,
+          family = family
         )
-        # for each peak, select 200 at random with matching GC content and accessibility
-        # sample at random from peaks on a different chromosome to the gene
-        peaks.test <- colnames(x = peak.access)
-        trans.peaks <- all.peaks[
-          !grepl(pattern = paste0("^", gene.chrom), x = all.peaks)
-        ]
-        meta.use <- meta.features[trans.peaks, ]
-        meta.use <- rbind(meta.use, meta.features[peaks.test, ])
-        bg.peaks <- lapply(
-          X = peaks.test,
-          FUN = MatchRegionStats,
-          meta.feature = meta.use,
-          features.match = c("count"),
-          n = n_sample,
-          verbose = FALSE
-        )
-        # run background correlations
-        pvals <- vector(mode = "numeric", length = length(x = peaks.test))
-        for (j in seq_along(along.with = peaks.test)) {
-          bg.access <- t(x = peak.data[unlist(x = bg.peaks[[j]]), ])
-          bg.coef <- cor(
-            x = as.matrix(x = bg.access),
-            y = as.matrix(x = gene.expression),
-            method = "spearman"
-          )
-          p <- ifelse(
-            test = coef.result[[j]] > 0,
-            yes = sum(bg.coef > coef.result[[j]]),
-            no = sum(bg.coef < coef.result[[j]])
-          ) / n_sample
-          pvals[[j]] <- p
+        lambda <- cvfit$lambda.1se
+        coef.results <- coef(object = cvfit, s = lambda)
+        coef.results <- coef.results[2:nrow(x = coef.results), ]
+        if (keep.all) {
+          # keep all genes so we know what was tested
+          coef.results.filtered <- coef.results
+        } else {
+          coef.results.filtered <- coef.results[coef.results != 0]
         }
-        names(x = coef.result) <- peaks.test
-        names(x = pvals) <- peaks.test
-        pval.vec <- c(pval.vec, pvals)
-        gene.vec <- c(gene.vec, rep(i, length(x = coef.result)))
-        coef.vec <- c(coef.vec, coef.result)
+        if (length(x = coef.results.filtered) == 0) {
+          return(list("gene" = NULL, "coef" = NULL, "pval" = NULL))
+        } else {
+          # compute null distribution with same lambda
+          # sample from peaks on a different chromosome to the gene
+          trans.peaks <- all.peaks[
+            !grepl(pattern = paste0("^", gene.chrom), x = all.peaks)
+          ]
+          peaks.test <- colnames(x = peak.access)
+          meta.use <- meta.features[trans.peaks, ]
+          meta.use <- rbind(meta.use, meta.features[peaks.test, ])
+          # sample matching total accessibility and GC
+          peak.random <- MatchRegionStats(
+            meta.feature = meta.use,
+            regions = peaks.test,
+            features.match = features.match,
+            n = n_sample,
+            verbose = FALSE
+          )
+          peak.access <- t(x = peak.data[peak.random, ])
+          cvfit <- glmnet(
+            x = peak.access,
+            y = gene.expression,
+            alpha = alpha,
+            family = family
+          )
+          coef.results.rand <- coef(object = cvfit, s = lambda)
+          coef.results.rand <- coef.results.rand[
+            2:nrow(x = coef.results.rand),
+          ]
+          # number of random peaks with coefficient at least as extreme as obs
+          pvals <- sapply(
+            X = coef.results.filtered,
+            FUN = function(x) {
+              if (x < 0) {
+                sum(coef.results.rand < x)
+              } else {
+                sum(coef.results.rand > x)
+              }
+            })
+          pvals <- pvals / n_sample
+          pval.vec <- c(pval.vec, pvals)
+          gene.vec <- c(gene.vec, rep(i, length(x = coef.results.filtered)))
+          coef.vec <- c(coef.vec, coef.results.filtered)
+        }
       }
       return(list("gene" = gene.vec, "coef" = coef.vec, "pval" = pval.vec))
     }