+++ NEWS.md
@@ -11,6 +11,7 @@ for different groups of cells separately by setting the `group.by` parameter
 * Added `show.bulk` parameter to `CoveragePlot()` to plot accessibility of all cells combined
 * Added ability to remove `Fragment` objects and modify the file path for existing
 fragment objects ([#206](https://github.com/timoast/signac/issues/206))
+* Improved feature matching in `MatchRegionStats()` function when matching distribution of multiple features (eg, GC content and overall accessibility)
 
 Bug fixes: